open groupCard

+++ src/storage/group/groupCreate.ts
@@ -1,12 +1,21 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import { AppError } from '@utils/AppError';
+
 import { GROUP_COLLECTION } from '@storage/storegeConfig';
+
 import { groupsGetAll } from './groupsGetAll';
 
 export async function groupCreate(newGroup: string) {
   try{
-    const storagedGroups = await groupsGetAll();
+    const storedGroups = await groupsGetAll();
 
-    const storage = JSON.stringify([...storagedGroups, newGroup])
+    const groupAlreadyExists = storedGroups.includes(newGroup)
+
+    if(groupAlreadyExists) {
+      throw new AppError('Já existe um grupo cadastrado com esse nome.')
+    }
+
+    const storage = JSON.stringify([...storedGroups, newGroup])
 
     await AsyncStorage.setItem(GROUP_COLLECTION, storage)
